# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
  internal_server_handlers_subscription.Subscription:
    properties:
      city:
        description: City for weather updates
        type: string
      confirmed:
        description: Whether the subscription is confirmed
        type: boolean
      email:
        description: Email address
        type: string
      frequency:
        description: |-
          Frequency of updates
          Enum: hourly, daily
        type: string
    type: object
  weather.Weather:
    properties:
      description:
        description: Weather description
        type: string
      humidity:
        description: Current humidity percentage
        type: number
      temperature:
        description: Current temperature
        type: number
    type: object
info:
  contact: {}
  description: Weather API application that allows users to subscribe to weather updates
    for their city.
  title: Weather Forecast API
  version: 1.0.0
paths:
  /confirm/{token}:
    get:
      description: Confirms a subscription using the token sent in the confirmation
        email.
      parameters:
      - description: Confirmation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Subscription confirmed successfully
          schema:
            type: string
        "400":
          description: Invalid token
          schema:
            type: string
        "404":
          description: Token not found
          schema:
            type: string
      summary: Confirm email subscription
      tags:
      - subscription
  /subscribe:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Subscribes an email to weather updates for a specific city with
        the given frequency.
      parameters:
      - description: Email address to subscribe
        in: formData
        name: email
        required: true
        type: string
      - description: City for weather updates
        in: formData
        name: city
        required: true
        type: string
      - description: Frequency of updates (hourly or daily)
        enum:
        - hourly
        - daily
        in: formData
        name: frequency
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Subscription successful. Confirmation email sent.
          schema:
            $ref: '#/definitions/internal_server_handlers_subscription.Subscription'
        "400":
          description: Invalid input
          schema:
            type: string
        "409":
          description: Email already subscribed
          schema:
            type: string
        "500":
          description: Internal api error
          schema:
            type: string
      summary: Subscribe to weather updates
      tags:
      - subscription
  /unsubscribe/{token}:
    get:
      description: Unsubscribes an email from weather updates using the token sent
        in emails.
      parameters:
      - description: Unsubscribe token
        in: path
        name: token
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Unsubscribed successfully
          schema:
            type: string
        "400":
          description: Invalid token
          schema:
            type: string
        "404":
          description: Token not found
          schema:
            type: string
      summary: Unsubscribe from weather updates
      tags:
      - subscription
  /weather:
    get:
      consumes:
      - application/json
      description: Returns the current weather forecast for the specified city using
        WeatherAPI.com.
      parameters:
      - description: City name for weather forecast
        in: query
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation - current weather forecast returned
          schema:
            $ref: '#/definitions/weather.Weather'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: City not found
          schema:
            type: string
      summary: Get current weather for a city
      tags:
      - weather
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Weather forecast operations
  name: weather
- description: Subscription management operations
  name: subscription